+++ sdk/resourcemanager/monitor/armmonitor/ze_generated_example_actiongroups_client_test.go
@@ -372,7 +372,7 @@ func ExampleActionGroupsClient_GetTestNotifications() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/monitor/resource-manager/Microsoft.Insights/stable/2021-09-01/examples/listActionGroups.json
-func ExampleActionGroupsClient_ListByResourceGroup() {
+func ExampleActionGroupsClient_NewListByResourceGroupPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -384,7 +384,7 @@ func ExampleActionGroupsClient_ListByResourceGroup() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListByResourceGroup("<resource-group-name>",
+	pager := client.NewListByResourceGroupPager("<resource-group-name>",
 		nil)
 	for pager.More() {
 		nextResult, err := pager.NextPage(ctx)